% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glmSparse}
\alias{glmSparse}
\title{Fitting sparse Generalized Linear Models}
\usage{
glmSparse(
  formula = NULL,
  family = stats::gaussian,
  data = NULL,
  weights = NULL,
  subset = NULL,
  na.action = stats::na.fail,
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = NULL,
  doFit = TRUE,
  control = list(...),
  model = TRUE,
  contrasts = NULL,
  x = NULL,
  y = NULL,
  drop.unused.levels = FALSE,
  ...
)
}
\arguments{
\item{formula}{An object of class \link{formula}, (or one that can be
coerced to that class): a symbolic description of the model to be fitted.}

\item{family}{A description of the error distribution and link function to be
used in the model. This can be a character string naming a family function,
a family function or the result of a call to a family function. (See
\link{family} for details of family functions.)}

\item{data}{An optional data frame, list or environment (or object coercible
by \link{as.data.frame} to a data frame) containing the variables in the
model.}

\item{weights}{an optional vector of ‘prior weights’ to be used in the
fitting process. Should be NULL or a numeric vector.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{A function which indicates what should happen when the data
contain \code{NA}s. The default is set by the na.action setting of \link{options},
and is \link{na.fail} if that is unset. The ‘factory-fresh’ default is
\link{na.omit}. Another possible value is \code{NULL}, no action. Value
\link{na.exclude} can be useful.}

\item{start}{Starting values for the parameters in the linear predictor.}

\item{etastart}{Starting values for the parameters in the predictor itself.}

\item{mustart}{Starting values for the parameters in the vector of means.}

\item{offset}{This can be used to specify an \emph{a priori} known component
to be included in the linear predictor during fitting. This should be \code{NULL}
or a numeric vector of length equal to the number of cases. One or more
offset terms can be included in the formula instead or as well, and if more
than one is specified, their sum is used. See \link{model.offset}.}

\item{doFit}{Logical indicating if the model should be fitted
(or just returned unfitted).}

\item{control}{A list with options on fitting; currently passed unchanged to
(hidden) function \code{IRLS()}.}

\item{model}{Currently ignored; here for back-compatibility with \link{glm}.}

\item{contrasts}{Currently ignored; here for back-compatibility with \link{glm}.}

\item{x}{Alternate interface for modeling. An already-constructed sparse
model matrix (class 'dgCMatrix').}

\item{y}{Alternate interface for modeling. A numeric response vector.}

\item{drop.unused.levels}{Should factors have unused levels dropped? Defaults
to \code{FALSE}.}

\item{...}{Potentially arguments passed on to fitter functions; currently
unused.}
}
\value{
An object of class \code{glpModel}.
}
\description{
\code{glmSparse} is a thin wrapper around \code{MatrixModels::glm4}. The notable
differences are that \code{glmSparse} always fits using a sparse model matrix,
and that it provides an alternate interface for situations where the user
wants to directly provide a sparse model matrix and response variable instead
of using the more standard formula interface. For more details on the
underlying implementation, see \code{MatrixModels::glm4}.
}
\seealso{
\code{\link[Matrix:sparse.model.matrix]{Matrix::sparse.model.matrix()}} for information on creating sparse
model matrices, and \code{\link[MatrixModels:glm4]{MatrixModels::glm4()}} to understand more details on
the implementation.
}
