% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glm4}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:reexports]{tidy}}}

  \item{MatrixModels}{\code{\link[MatrixModels]{glm4}}}
}}

